Moved AsyncTask to its own top level file.  This solves the potential memory leak problem, and is better code organization for including search in other activities.
