+++ app/src/main/java/com/weebly/stevelosk/sewingpatternapp/BasicSearchActivity.java
@@ -83,7 +83,7 @@ public class BasicSearchActivity extends AppCompatActivity {
             public void onClick(View view) {
                 AsyncTask searchTask = new AsyncSearchTask();
                 Object[] params = {getApplicationContext(),
-                        searchEditText.getText().toString(), patterns};
+                        searchEditText.getText().toString(), patterns, db, pa};
                 searchTask.execute(params);
 
             }
@@ -118,7 +118,7 @@ public class BasicSearchActivity extends AppCompatActivity {
                         }
                         searchTask = new AsyncSearchTask();
                         Object[] params = {getApplicationContext(),
-                                searchEditText.getText().toString(), patterns};
+                                searchEditText.getText().toString(), patterns, db, pa};
                         searchTask.execute(params);
 
                     }
@@ -126,84 +126,6 @@ public class BasicSearchActivity extends AppCompatActivity {
 
             }
         });
-    }
-
-
-    class AsyncSearchTask extends AsyncTask<Object, Void, Integer> {
-
-        private static final int ASYNC_TASK_CANCELLED = -1;
-        private static final int ASYNC_TASK_COMPLETED = 0;
-
-        @Override
-        protected Integer doInBackground(Object... objects) {
-
-            // create database connection
-            Context ctx = (Context) objects[0];
-
-            // get query content and result set target from packaged objects
-            String searchStr = (String) objects[1];
-            ArrayList<Pattern> patterns = (ArrayList<Pattern>) objects[2];
-
-            try {
-
-                // clear the old result list
-                patterns.clear();
-
-                // query
-                String[] predicate = new String[2];
-                // surround search string with "%" so partial matches can be found
-                predicate[0] = "%" + searchStr + "%";
-                predicate[1] = "%" + searchStr + "%";
-                Cursor cursor = db.getPatternBy_ID_OR_Content(predicate);
-                if (this.isCancelled()) {
-                    return ASYNC_TASK_CANCELLED;
-                }
-
-                // populate the results from the query result set
-                while (cursor.moveToNext()) {
-                    Pattern p = PatternDBAdapter.getPatternFromCursor(cursor);
-                    patterns.add(p);
-                    if (this.isCancelled()) {
-                        return ASYNC_TASK_CANCELLED;
-                    }
-                }
-
-                return ASYNC_TASK_COMPLETED;
-            } catch (SQLException e) {
-                Toast.makeText(ctx, e.getMessage(), Toast.LENGTH_LONG).show();
-
-                return 1;
-            }
-        }
-
-        @Override
-        protected void onPostExecute(Integer result) {
-
-            if (result == ASYNC_TASK_COMPLETED) {
-                pa.notifyDataSetChanged();
-               /* if (patterns.size() == 0) {
-                    Toast.makeText(getApplicationContext(), R.string.noPatternsFound,
-                            Toast.LENGTH_SHORT).show();
-                } */
-            } else if (result > ASYNC_TASK_COMPLETED) {
-                Toast.makeText(getApplicationContext(), R.string.database_error,
-                        Toast.LENGTH_LONG).show();
-            }
-
-            // Make sure AsyncTask dies
-            try {
-                this.finalize();
-
-            } catch (Throwable throwable) {
-                throwable.printStackTrace();
-            }
-
-        }
-
-        @Override
-        protected void onCancelled() {
-            super.onCancelled();
-        }
     }
 
     @Override
